requirements.txt
# requirements.txt

# Core Frameworks
fastapi
uvicorn[standard]
pyside6
httpx
pydantic # Pydantic V2 is critical

# LLM & RAG
# ollama # This is usually a CLI tool, not a Python library to list here.
# We interact with Ollama via HTTP requests using httpx.

# Vector Store (FAISS)
faiss-cpu # Or faiss-gpu if you have the hardware and setup for it
numpy # Often a dependency for FAISS and numerical operations

# Document Loading (add more as needed)
python-dotenv # For loading .env files (optional, but good practice)
# PyPDF2 # Example if you plan to support PDF files
# python-docx # Example for .docx files
# beautifulsoup4 # Example for HTML files
# unstructured # A comprehensive library for various unstructured data

# Logging & Utilities
# Standard library: logging, pathlib, sys, os, datetime, uuid

# For development and testing (optional to list here, can be in a dev-requirements.txt)
# pytest
# pytest-asyncio